Move direct-wallet plugin under integrations/deposits/
Groups deposit-method plugins under a single parent directory so future
methods (pull-allowance, one-time-address) can live alongside.
